+++ apps/passport-server/src/services/generic-issuance/pipelines/types.ts
@@ -1,0 +1,17 @@
+import { BasePipelineCapability, PipelineType } from "../types";
+import { LemonadePipeline } from "./LemonadePipeline";
+import { PretixPipeline } from "./PretixPipeline";
+
+/**
+ * Each new type of {@link Pipeline} needs to be added to this type
+ * declaration.
+ */
+export type Pipeline = LemonadePipeline | PretixPipeline;
+
+/**
+ * Interface from which all {@link Pipeline}s derive.
+ */
+export interface BasePipeline {
+  type: PipelineType;
+  capabilities: readonly BasePipelineCapability[];
+}
